requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.8.0
tqdm>=4.66.0
trimesh>=4.2.0
pyrender>=0.1.45
open3d>=0.17.0
opencv-python>=4.8.0.76
transformers>=4.36.0
huggingface-hub>=0.19.0
accelerate>=0.24.0
safetensors>=0.4.0
pyyaml>=6.0.1
pandas>=2.1.0
imageio>=2.34.0
